pyproject.toml
[project]
name = "medical-image-std"
version = "0.2.0"
description = "A standardized Python framework for medical image processing."
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}

authors = [
    {name = "Hamza Gbada", email = "hamza.gbada@gmail.com"},
]

dependencies = [
    "pydicom>=2.4.4",
    "numpy>=1.26.4",
    "Pillow>=10.3.0",
    "requests>=2.28.1",
    "matplotlib",
    "scipy",
    "scikit-image",
    "torch",
    "torchvision",
    "transformers",
    "pandas",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "ruff", "mypy"]
gpu = ["torch", "torchvision"]
all = ["medical-image-std[dev,gpu]"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["medical_image*"]

[tool.setuptools.package-data]
medical_image = ["py.typed"]

# ---------------------
# uv configuration
# ---------------------
[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
